feat: add settings screen and enable light theme
