+++ AppEstoque/app/src/main/AndroidManifest.xml
@@ -36,6 +36,9 @@
             android:screenOrientation="landscape"/>
         <activity android:name=".checklist.ChecklistExpedicaoActivity"
             android:screenOrientation="landscape"/>
+        <activity android:name=".SettingsActivity"
+            android:exported="false"
+            android:screenOrientation="landscape"/>
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="${applicationId}.fileprovider"
