+++ AppEstoque/app/src/main/java/com/example/apestoque/MainActivity.kt
@@ -1,11 +1,9 @@
 package com.example.apestoque
 
-import android.content.Context
+import android.content.Intent
 import android.os.Bundle
-import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import androidx.core.widget.doAfterTextChanged
 import androidx.fragment.app.Fragment
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import androidx.viewpager2.widget.ViewPager2
@@ -17,6 +15,8 @@ import com.example.apestoque.fragments.LogisticaFragment
 import com.example.apestoque.fragments.InspecionarFragment
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
+import android.view.Menu
+import android.view.MenuItem
 
 class MainActivity : AppCompatActivity() {
 
@@ -26,20 +26,6 @@ class MainActivity : AppCompatActivity() {
 
         // Toolbar
         findViewById<Toolbar>(R.id.toolbar).also { setSupportActionBar(it) }
-
-        // Operador de Suprimentos
-        val prefs = getSharedPreferences("app", Context.MODE_PRIVATE)
-        val etOperador = findViewById<EditText>(R.id.etOperadorSuprimentos)
-        etOperador.setText(prefs.getString("operador_suprimentos", ""))
-        etOperador.doAfterTextChanged {
-            prefs.edit().putString("operador_suprimentos", it.toString()).apply()
-        }
-
-        val etApiIp = findViewById<EditText>(R.id.etApiIp)
-        etApiIp.setText(prefs.getString("api_ip", "192.168.0.135"))
-        etApiIp.doAfterTextChanged {
-            prefs.edit().putString("api_ip", it.toString()).apply()
-        }
 
         // Views
         val swipe = findViewById<SwipeRefreshLayout>(R.id.swipeRefresh)
@@ -69,6 +55,21 @@ class MainActivity : AppCompatActivity() {
         swipe.setOnRefreshListener {
             pager.adapter?.notifyItemChanged(pager.currentItem)
             swipe.isRefreshing = false
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.menu_main, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            R.id.action_settings -> {
+                startActivity(Intent(this, SettingsActivity::class.java))
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
         }
     }
 }
