+++ AppEstoque/app/src/main/res/layout/activity_main.xml
@@ -29,24 +29,6 @@
                     android:textColor="@android:color/white"
                     android:textSize="20sp"/>
 
-                <EditText
-                    android:id="@+id/etOperadorSuprimentos"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="4dp"
-                    android:background="@android:color/white"
-                    android:hint="Nome do operador Sumprimetos"
-                    android:maxLines="1"/>
-
-                <EditText
-                    android:id="@+id/etApiIp"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="4dp"
-                    android:background="@android:color/white"
-                    android:hint="IP da API"
-                    android:maxLines="1"/>
-
             </LinearLayout>
         </androidx.appcompat.widget.Toolbar>
