+++ AppEstoque/app/src/main/res/layout/activity_settings.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="16dp">
+
+    <EditText
+        android:id="@+id/etOperadorSuprimentos"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:hint="Nome do operador Sumprimetos"
+        android:maxLines="1"/>
+
+    <EditText
+        android:id="@+id/etApiIp"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:hint="IP da API"
+        android:maxLines="1"/>
+
+</LinearLayout>
