+++ AppEstoque/app/src/main/res/values/themes.xml
@@ -1,23 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
 
-    <!-- Tema principal sem ActionBar -->
-    <style name="Theme.AppEstoque" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <!-- Tema principal sempre claro sem ActionBar -->
+    <style name="Theme.AppEstoque" parent="Theme.MaterialComponents.Light.NoActionBar">
         <item name="colorPrimary">@color/purple_500</item>
         <item name="colorOnPrimary">@color/white</item>
         <item name="android:statusBarColor">@color/purple_700</item>
         <item name="colorSecondary">@color/teal_200</item>
         <item name="colorOnSecondary">@color/black</item>
         <item name="android:windowBackground">?attr/colorSurface</item>
-    </style>
-
-    <!-- Variante clara -->
-    <style name="Theme.AppEstoque.Light" parent="Theme.MaterialComponents.Light.NoActionBar">
-        <item name="colorPrimary">@color/purple_200</item>
-        <item name="colorOnPrimary">@color/black</item>
-        <item name="android:statusBarColor">@color/purple_500</item>
-        <item name="colorSecondary">@color/teal_700</item>
-        <item name="colorOnSecondary">@color/white</item>
     </style>
 
 </resources>
